backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
email-validator==2.2.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
